Move OpenSearch stuff from `AstExpressionBuilder`s.
